backend/pyproject.toml
[project]
name = "slide-deck-generator"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "anthropic>=0.42.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "websockets>=13.0",
    "pypdf>=4.0.0",
]
